# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bakery Service API",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/orders": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Membuat Pesanan [Idempoten - Pola Stripe]",
        "description": "**[Operasi Idempoten - Pola Stripe]**\n\nMembuat transaksi pesanan baru dan melakukan alokasi pengurangan stok produk secara otomatis.\n\nEndpoint ini mewajibkan/mendukung header `Idempotency-Key` (UUID v4) guna menjamin tidak terjadinya duplikasi order dan pengurangan stok berganda apabila terjadi pengulangan permintaan dari klien. Pengulangan permintaan dengan kunci sama dan payload sama akan mengembalikan respons *cache* (dengan header `Idempotent-Replay: true` dan `X-Cache-Lookup: HIT`). Jika kunci yang sama dipakai dengan payload berbeda, server mengembalikan status `HTTP 409 Conflict`.",
        "operationId": "create_order_v1_orders_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Kunci unik (UUID v4) untuk mencegah duplikasi pemrosesan pesanan",
              "title": "Idempotency-Key"
            },
            "description": "Kunci unik (UUID v4) untuk mencegah duplikasi pemrosesan pesanan"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pesanan berhasil dibuat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parameter tidak valid atau stok tidak mencukupi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pelanggan atau Produk tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Konflik Idempotency-Key digunakan dengan parameter berbeda",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Daftar Pesanan",
        "description": "Mengambil daftar transaksi pesanan dengan dukungan navigasi kursor (cursor-based pagination) dan filter status.",
        "operationId": "list_orders_v1_orders_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Jumlah data per halaman (1–100)",
              "examples": [
                10
              ],
              "default": 10,
              "title": "Limit"
            },
            "description": "Jumlah data per halaman (1–100)"
          },
          {
            "name": "starting_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID pesanan sebagai kursor untuk mengambil data setelah entri ini",
              "examples": [
                "ord_20260925_001"
              ],
              "title": "Starting After"
            },
            "description": "ID pesanan sebagai kursor untuk mengambil data setelah entri ini"
          },
          {
            "name": "ending_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID pesanan sebagai kursor untuk mengambil data sebelum entri ini",
              "title": "Ending Before"
            },
            "description": "ID pesanan sebagai kursor untuk mengambil data sebelum entri ini"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter berdasarkan status pesanan (misal: pending_payment, paid, completed)",
              "title": "Status"
            },
            "description": "Filter berdasarkan status pesanan (misal: pending_payment, paid, completed)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/orders/{order_id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Detail Pesanan",
        "description": "Mengambil rincian data transaksi pesanan berdasarkan ID pesanan.",
        "operationId": "get_order_v1_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID unik pesanan",
              "examples": [
                "ord_20260925_001"
              ],
              "title": "Order Id"
            },
            "description": "ID unik pesanan"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pesanan tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/orders/{order_id}/cancel": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Membatalkan Pesanan [Idempoten - State Transition]",
        "description": "**[Operasi Idempoten - State Transition & Idempotency-Key]**\n\nMembatalkan pesanan yang belum selesai serta mengembalikan alokasi kuota stok produk.\nMendukung header `Idempotency-Key` dan transisi status aman sehingga pemanggilan berulang tidak akan mengembalikan stok roti lebih dari 1 kali.",
        "operationId": "cancel_order_v1_orders__order_id__cancel_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID unik pesanan yang akan dibatalkan",
              "examples": [
                "ord_20260925_002"
              ],
              "title": "Order Id"
            },
            "description": "ID unik pesanan yang akan dibatalkan"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Kunci unik UUID v4",
              "title": "Idempotency-Key"
            },
            "description": "Kunci unik UUID v4"
          }
        ],
        "responses": {
          "200": {
            "description": "Pesanan berhasil dibatalkan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Status pesanan tidak memenuhi syarat untuk dibatalkan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pesanan tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Konflik kunci idempotensi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "examples": [
              "invalid_request_error"
            ]
          },
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "idempotency_conflict"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "Keys for idempotent requests cannot be reused with different request body."
            ]
          },
          "param": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Param",
            "examples": [
              "Idempotency-Key"
            ]
          }
        },
        "type": "object",
        "required": [
          "type",
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OrderCreate": {
        "properties": {
          "customer_id": {
            "type": "string",
            "title": "Customer Id",
            "examples": [
              "cus_budi01"
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemInput"
            },
            "type": "array",
            "title": "Items"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "examples": [
              "Potong slice tebal"
            ]
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "items"
        ],
        "title": "OrderCreate"
      },
      "OrderItemDetail": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "unit_price": {
            "type": "integer",
            "title": "Unit Price"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "subtotal": {
            "type": "integer",
            "title": "Subtotal"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "name",
          "unit_price",
          "quantity",
          "subtotal"
        ],
        "title": "OrderItemDetail"
      },
      "OrderItemInput": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id",
            "examples": [
              "prod_sourdough01"
            ]
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "examples": [
              2
            ]
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "quantity"
        ],
        "title": "OrderItemInput"
      },
      "OrderListResponse": {
        "properties": {
          "object": {
            "type": "string",
            "title": "Object",
            "default": "list",
            "examples": [
              "list"
            ]
          },
          "url": {
            "type": "string",
            "title": "Url",
            "default": "/v1/orders",
            "examples": [
              "/v1/orders"
            ]
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "examples": [
              false
            ]
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/OrderResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "has_more",
          "data"
        ],
        "title": "OrderListResponse"
      },
      "OrderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "ord_20260925_001"
            ]
          },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "order",
            "examples": [
              "order"
            ]
          },
          "customer": {
            "type": "string",
            "title": "Customer",
            "examples": [
              "cus_budi01"
            ]
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Status pesanan: pending_payment, paid, baking, ready_for_pickup, completed, cancelled.",
            "examples": [
              "pending_payment"
            ]
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "idr",
            "examples": [
              "idr"
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemDetail"
            },
            "type": "array",
            "title": "Items"
          },
          "total_amount": {
            "type": "integer",
            "title": "Total Amount",
            "examples": [
              90000
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer",
          "status",
          "items",
          "total_amount"
        ],
        "title": "OrderResponse"
      },
...
      "RFCErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "RFCErrorResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}